server/requirements.txt
google-cloud-bigquery==3.3.5
fastapi==0.85.1
pandas==1.5.0
pandas-gbq==0.17.9
protobuf==4.21.8
uvicorn==0.11.3